+++ app/src/main/java/com/sopt/now/network/request/RequestChPwDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.network
+package com.sopt.now.network.request
 
 import kotlinx.serialization.Serializable
