+++ app/src/main/java/com/sopt/now/network/request/RequestLoginDto.kt
@@ -1,11 +1,9 @@
-package com.sopt.now.network
+package com.sopt.now.network.request
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
+data class RequestLoginDto(
     val authenticationId: String,
     val password: String,
-    val nickname: String,
-    val phone: String,
 )
