+++ app/src/main/java/com/sopt/now/network/request/RequestSignUpDto.kt
@@ -1,9 +1,11 @@
-package com.sopt.now.network
+package com.sopt.now.network.request
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestLoginDto(
+data class RequestSignUpDto(
     val authenticationId: String,
     val password: String,
+    val nickname: String,
+    val phone: String,
 )
