+++ app/src/main/java/com/sopt/now/network/response/ResponseDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.network
+package com.sopt.now.network.response
 
 import kotlinx.serialization.Serializable
