+++ app/src/main/java/com/sopt/now/network/service/AuthService.kt
@@ -1,5 +1,9 @@
-package com.sopt.now.network
+package com.sopt.now.network.service
 
+import com.sopt.now.network.request.RequestChPwDto
+import com.sopt.now.network.request.RequestLoginDto
+import com.sopt.now.network.request.RequestSignUpDto
+import com.sopt.now.network.response.ResponseDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.PATCH
