+++ app/src/main/java/com/sopt/now/network/service/ServicePool.kt
@@ -1,0 +1,9 @@
+package com.sopt.now.network.service
+
+import com.sopt.now.network.ApiFactory
+import com.sopt.now.network.FollowerApiFactory
+
+object ServicePool {
+    val authService = ApiFactory.create<AuthService>()
+    val followerService = FollowerApiFactory.create<FollowerService>()
+}
