+++ app/src/main/java/com/sopt/now/ui/signup/SignupViewModel.kt
@@ -5,9 +5,9 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.sopt.now.data.Sopt
 import com.sopt.now.data.User
-import com.sopt.now.network.RequestSignUpDto
-import com.sopt.now.network.ResponseDto
-import com.sopt.now.network.ServicePool
+import com.sopt.now.network.request.RequestSignUpDto
+import com.sopt.now.network.response.ResponseDto
+import com.sopt.now.network.service.ServicePool
 import com.sopt.now.ui.AuthState
 import org.json.JSONObject
 import retrofit2.Call
